# codr-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Users demo
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
paths:
...
  /users/current/comments:
    x-swagger-router-controller: comments
    get:
      description: Return all comments that the current user can see.
      operationId: getComments
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/CommentResponseList'
        "401":
          description: Not allowed
...
  /modules/{module_id}/courses/{course_id}/exercises/{exercise_id}/projects/{project_id}/files/{file_id}/comments:
    x-swagger-router-controller: comments
    post:
      description: Creates a comment
      operationId: postComment
      parameters:
        - name: module_id
          in: path
          description: ID of the module from which to fetch courses.
          required: true
          type: integer
          format: int64
        - name: course_id
          in: path
          description: ID of the course from which to fetch exercises.
          required: true
          type: integer
          format: int64
        - name: exercise_id
          in: path
          description: ID of the exercise to fetch.
          required: true
          type: integer
          format: int64
        - name: project_id
          in: path
          description: ID of the project to fetch.
          required: true
          type: integer
          format: int64
        - name: file_id
          in: path
          description: ID of file to fetch comments for
          required: true
          type: integer
          format: int64
        - name: comment
          in: body
          required: true
          schema:
            type: object
            properties:
              parent_id: 
                type: integer
                format: int64
              line_range:
                type: string
              content:
                type: string
      responses:
        "201":
          description: Success, new comment has been created.
        "401":
          description: Error, user is not allowed to access this page.
        "404":
          description: Not found.
        "405":
          description: Error, invalid input.
        "500":
          description: Internal server error
  /modules/{module_id}/courses/{course_id}/exercises/{exercise_id}/projects/{project_id}/files/{file_id}/comments/{comment_id}:
    x-swagger-router-controller: comments
    put:
      description: Updates the specified comment.
      operationId: putCommentById
      parameters:
        - name: module_id
          in: path
          description: ID of the module from which to fetch courses.
          required: true
          type: integer
          format: int64
        - name: course_id
          in: path
          description: ID of the course from which to fetch exercises.
          required: true
          type: integer
          format: int64
        - name: exercise_id
          in: path
          description: ID of the exercise to fetch.
          required: true
          type: integer
          format: int64
        - name: project_id
          in: path
          description: ID of the project to fetch.
          required: true
          type: integer
          format: int64
        - name: file_id
          in: path
          description: ID of file to fetch comments for
          required: true
          type: integer
          format: int64
        - name: comment_id
          in: path
          description: ID of comment
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: body of the request
          schema:
            type: object
            properties:
              content:
                type: string
              visible:
                type: boolean
      responses:
        "200":
          description: Success
        "401":
          description: Not allowed
        "404":
          description: Id not found
        "500":
          description: Internal server error
    delete:
      description: Deletes the specified comment.
      operationId: deleteCommentById
      parameters:
        - name: module_id
          in: path
          description: ID of the module from which to fetch courses.
          required: true
          type: integer
          format: int64
        - name: course_id
          in: path
          description: ID of the course from which to fetch exercises.
          required: true
          type: integer
          format: int64
        - name: exercise_id
          in: path
          description: ID of the exercise to fetch.
          required: true
          type: integer
          format: int64
        - name: project_id
          in: path
          description: ID of the project to fetch.
          required: true
          type: integer
          format: int64
        - name: file_id
          in: path
          description: ID of file to fetch comments for
          required: true
          type: integer
          format: int64
        - name: comment_id
          in: path
          description: ID of comment
          required: true
          type: integer
          format: int64
      responses:
        "204":
          description: Item deleted
        "401":
          description: Not allowed
        "404":
          description: Id not found
        "500":
          description: Internal server error
...
definitions:
...
  CommentResponseHead:
    type: object
    properties:
      id:
        type: integer
        format: int64
      parent_id:
        type: integer
        format: int64
      file_id:
        type: integer
        format: int64
      project_id:
        type: integer
        format: int64
      exercise_id:
        type: integer
        format: int64
      course_id:
        type: integer
        format: int64
      module_id:
        type: integer
        format: int64
      line_range:
        type: string
      content:
        type: string
      author_id:
        type: integer
        format: int64
      author_name:
        type: string
      visible:
        type: boolean
      timestamp:
        type: integer
        format: int64
      deleted:
        type: boolean
      can_edit:
        type: boolean
      can_delete:
        type: boolean
      can_toggle_visibility:
        type: boolean
      children:
        type: array
        items:
          $ref: '#/definitions/CommentResponseTail'
  CommentResponseTail:
    type: object
    properties:
      id:
        type: integer
        format: int64
      parent_id:
        type: integer
        format: int64
      file_id:
        type: integer
        format: int64
      project_id:
        type: integer
        format: int64
      exercise_id:
        type: integer
        format: int64
      course_id:
        type: integer
        format: int64
      module_id:
        type: integer
        format: int64
      line_range:
        type: string
      content:
        type: string
      author_id:
        type: integer
        format: int64
      author_name:
        type: string
      visible:
        type: boolean
      timestamp:
        type: integer
        format: int64
      deleted:
        type: boolean
      can_edit:
        type: boolean
      can_delete:
        type: boolean
      can_toggle_visibility:
        type: boolean
      children:
        type: array
        items:
          type: object
  CommentResponseList:
      type: array
      items:
        $ref: '#/definitions/CommentResponseHead'
...